requirements.txt
appdirs==1.4.3
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.3
django-environ==0.4.5
django-on-heroku==1.1.2
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
progress==1.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
pyparsing==2.4.6
pytest==6.1.2
pytest-django==4.1.0
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
appdirs==1.4.3
appnope==0.1.2
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
boto3==1.17.73
botocore==1.20.73
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.3
django-environ==0.4.5
django-on-heroku==1.1.2
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
progress==1.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
pyparsing==2.4.6
pytest==6.1.2
pytest-django==4.1.0
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
s3transfer==0.4.2
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
